Move Source classes to new source.py module and define them via entry-points in data.py, also fix several issues due to these changes
