+++ pyproject.toml
@@ -73,8 +73,11 @@ swan = "rompy.swan.config:SwanConfig"
 swan_components = "rompy.swan.config:SwanConfigComponents"
 schism = "rompy.schism.config:SchismCSIROConfig"
 
-[project.entry-points."rompy.data_source"]
-
+[project.entry-points."rompy.source"]
+dataset = "rompy.core.source:SourceDataset"
+file = "rompy.core.source:SourceFile"
+intake = "rompy.core.source:SourceIntake"
+datamesh = "rompy.core.source:SourceDatamesh"
 
 [project.entry-points."intake.catalogs"]
 "rompy_data" = "rompy:cat"
