+++ rompy/core/__init__.py
@@ -1,12 +1,5 @@
 from .config import BaseConfig
-from .data import (
-    DataBlob,
-    DataGrid,
-    SourceDataset,
-    SourceFile,
-    SourceIntake,
-    SourceDatamesh,
-)
+from .data import DataBlob, DataGrid
 from .filters import *
 from .grid import BaseGrid, RegularGrid
 from .time import TimeRange
