+++ rompy/core/boundary.py
@@ -7,16 +7,10 @@ from typing import Literal, Optional, Union
 import numpy as np
 import wavespectra
 import xarray as xr
-from pydantic import Field, field_validator, model_validator
+from pydantic import Field, field_validator
 
-from rompy.core.data import (
-    DataGrid,
-    SourceBase,
-    SourceDatamesh,
-    SourceDataset,
-    SourceFile,
-    SourceIntake,
-)
+from rompy.core.source import SourceBase
+from rompy.core.data import DataGrid
 from rompy.core.grid import RegularGrid
 from rompy.core.time import TimeRange
 from rompy.settings import BOUNDARY_SOURCE_TYPES, SPEC_BOUNDARY_SOURCE_TYPES
