+++ rompy/settings.py
@@ -5,7 +5,7 @@ import os
 # working in a python environement
 DATA_SOURCE_TYPES = os.getenv(
     "DATA_SOURCE_TYPES",
-    "rompy.core.data.SourceDataset,rompy.core.data.SourceFile,rompy.core.data.SourceIntake,rompy.core.data.SourceDatamesh",
+    "rompy.core.source.SourceDataset,rompy.core.source.SourceFile,rompy.core.source.SourceIntake,rompy.core.source.SourceDatamesh",
 )
 
 BOUNDARY_SOURCE_TYPES = os.getenv("BOUNDARY_SOURCE_TYPES", DATA_SOURCE_TYPES)
