+++ rompy/utils.py
@@ -301,6 +301,7 @@ def process_setting(settings_str: str):
         else:
             return Union[tuple(data_source_types)]
     else:
+        import ipdb; ipdb.set_trace()
         raise ValueError("No valid dataset types specified")
