+++ tests/schism/utils.py
@@ -1,3 +1,4 @@
+import os
 import logging
 import tarfile
 from datetime import datetime
