+++ tests/test_swanboundary.py
@@ -5,7 +5,8 @@ from wavespectra import read_swan
 
 from rompy.core.time import TimeRange
 from rompy.swan.grid import SwanGrid
-from rompy.core.boundary import SourceFile, SourceIntake, SourceWavespectra
+from rompy.core.source import SourceFile, SourceIntake
+from rompy.core.boundary import SourceWavespectra
 from rompy.swan.boundary import Boundnest1, BoundspecSide, BoundspecSegmentXY
