+++ tests/test_swantemplate.py
@@ -1,13 +1,12 @@
 from datetime import datetime
 from pathlib import Path
-
 import numpy as np
 import pytest
 import xarray as xr
 from utils import compare_files
 
 from rompy.model import ModelRun
-from rompy.core.data import SourceFile
+from rompy.core.source import SourceFile
 from rompy.core import TimeRange
 from rompy.core.types import DatasetCoords
 from rompy.swan import Boundnest1, SwanConfig, SwanDataGrid, SwanGrid
